interaction/requirements.txt
# example:
pwntools
